feat: option item input의 type 분리하고 components로 이동
