+++ src/pages/menu/components/index.ts
@@ -3,3 +3,4 @@ export { default as LabelInput } from "./labelInput";
 export { default as LabelTextarea } from "./labelTextarea";
 export { default as OptionModal } from "./optionModal";
 export { default as OnOffButton } from "./onOffButton";
+export { default as OptionItemInput } from "./optionItemInput";
